+++ github-notification/src/main.rs
@@ -1,7 +1,9 @@
 use github_notification::env::{get_github_personal_access_token, get_slack_webhook_url_from_env};
 use github_notification::github::{get_my_issues, sort_issues};
 use github_notification::sentry::initialize_sentry;
-use github_notification::slack::{create_payload_for_slack, notify_by_slack};
+use github_notification::slack::{
+    notification::notify_by_slack, payload::create_payload_for_slack,
+};
 
 #[tokio::main]
 async fn main() {
