+++ github-notification/src/slack/mod.rs
@@ -1,0 +1,3 @@
+pub mod message;
+pub mod notification;
+pub mod payload;
